+++ src/components/Order/Order.module.scss
@@ -8,38 +8,38 @@
     display: flex;
     justify-content: center;
     margin-bottom: 80px;
+    
+    .orderContainer {
+        height: 100%;
+        position: absolute;
+        top: -50%;
+        background-color: $secondary;
+        padding: 70px 110px;
+        border-radius: $radius-primary;
+        height: auto;
+        
+        @media(max-width: 1200px) {
+            padding: 50px 60px;
+        }
+        
+        @media(max-width: 1000px) {
+            top: 0%;
+        }
 
-.orderContainer {
-    height: 100%;
-    position: absolute;
-    top: -50%;
-    background-color: $secondary;
-    padding: 70px 110px;
-    border-radius: $radius-primary;
-    height: auto;
+        @media(max-width: 1000px) {
+            padding-left: 20px;
+            padding-right: 20px;
+        }
 
-     @media(max-width: 1200px) {
-        padding: 50px 60px;
-     }
-
-     @media(max-width: 1000px) {
-        top: 0%;
-     }
-
-     @media(max-width: 1000px) {
-        padding-left: 20px;
-        padding-right: 20px;
-     }
-
-     @include media-query(md) {
-        padding-top: 40px;
-        padding-bottom: 40px;
-     }
-
-     @media(max-width: 550px) {
-        padding-left: 10px;
-        padding-right: 10px;
-     }
+        @include media-query(md) {
+            padding-top: 40px;
+            padding-bottom: 40px;
+        }
+        
+        @media(max-width: 550px) {
+            padding-left: 10px;
+            padding-right: 10px;
+        }
 
     .orderBlock {
         width: 100%;
@@ -65,17 +65,6 @@
             gap: 10px;
         }
 
-        button {
-        // width: 100%;
-        // height: 70px;
-        // padding-right: 20px;
-        // @include flex(row, flex-end, center, nowrap);
-        // @include media-query(sm) {
-        //     height: 56px;
-        // }
-
-    }
-
         .btnBlock {
             width: 100%;
             height: 100%;
@@ -90,24 +79,6 @@
                 margin-right: 0px;
             }
 
-            // .descriptionBtn {
-            //     @include typography(400, 20px, 19.5px, $text-secondary);
-            //     margin-bottom: 20px;
-
-            //     @include media-query(lg) {
-            //         margin-bottom: 10px;
-            //         font-size: 16px;
-            //     }
-
-            //     @include media-query(sm) {
-            //         margin-bottom: 5px;
-            //     }
-            // }
-
-            // .btnInfo {
-            //     background-color: $text-white;
-            // }
-
             .btnSearch {
                 background-color: $color-8;
                 justify-content: center;
@@ -120,5 +91,12 @@
             }
         }
     }
+}
+.error {
+    color: red;
+    z-index: 1;
+    position: absolute;
+    right: 110px;
+    margin-top: 20px;
 }
 }
